pyproject.toml
[project]
name = "ats-cv-search"
version = "0.1.0"
description = "ATS CV Search Application - Tubes 3 Stima"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "PyQt5>=5.15.0",
    "mysql-connector-python>=8.0.0",
    "PyPDF2>=3.0.0",
    "sqlalchemy>=2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build]
include = [
    "src/**/*.py",
    "README.md",
    "LICENSE",
]
exclude = [
    "*.pyc",
    "__pycache__",
    ".git",
    ".venv",
    "data/*",
    "doc/*",
]